[JT-57] refactor: 애플리케이션 비즈니스 로직과 도메인 비즈니스 로직 분리
- module-domain.application -> module-domain.service 패키지로 변경
- module.application에 application 패키지 생성
